pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-analytics-mcp"
version = "0.1.0"
description = "MCP server for Google Analytics 4 - setup properties, streams, and query reports via Admin & Data APIs"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Ing. Luca Stucchi", email = "luca.stucchi@gmail.com" }]
keywords = ["mcp", "google-analytics", "ga4", "analytics", "reporting"]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "mcp[cli]>=1.26.0",
    "google-analytics-admin>=0.25.0",
    "google-analytics-data>=0.20.0",
    "google-auth>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
]

[project.urls]
Homepage = "https://github.com/stucchi/google-analytics-mcp"
Repository = "https://github.com/stucchi/google-analytics-mcp"
Issues = "https://github.com/stucchi/google-analytics-mcp/issues"

[project.scripts]
google-analytics-mcp = "google_analytics_mcp.server:main"

[tool.setuptools.packages.find]
where = ["src"]
